Move some of Dropwizard and Consul related Code to separate project
