+++ lib/dropwizard/base/build.gradle
@@ -1,0 +1,13 @@
+apply from: rootProject.file('gradle/common_deps.gradle')
+def dropwizardVersion = "0.8.0"
+
+dependencies {
+    compile 'com.hubspot.dropwizard:dropwizard-guice:0.8.0'
+    compile 'com.netflix.governator:governator:1.3.3'
+
+    compile "io.dropwizard:dropwizard-core:$dropwizardVersion"
+    compile "io.dropwizard:dropwizard-client:$dropwizardVersion"
+    compile 'com.wikia:dropwizard-logstash-encoder:1.0.2'
+
+    testCompile "io.dropwizard:dropwizard-testing:$dropwizardVersion"
+}
