+++ lib/dropwizard/build.gradle
@@ -1,0 +1,3 @@
+subprojects {
+    group 'com.wikia.dropwizard'
+}
