+++ lib/dropwizard/consul/build.gradle
@@ -1,0 +1,8 @@
+apply from: rootProject.file('gradle/common_deps.gradle')
+dependencies {
+    compile project(':lib:dropwizard:base')
+    compile project(':lib:consul-config')
+    compile('com.orbitz.consul:consul-client:0.8') {
+        exclude group: 'org.apache.cxf'
+    }
+}
