+++ lib/dropwizard/consul/src/main/java/com/wikia/dropwizard/consul/bundle/ProvidesConsulConfiguration.java
@@ -1,4 +1,4 @@
-package com.wikia.pandora.core.consul;
+package com.wikia.dropwizard.consul.bundle;
 
 public interface ProvidesConsulConfiguration {
   public ConsulConfiguration getConsulConfiguration();
