+++ lib/dropwizard/consul/src/main/java/com/wikia/dropwizard/consul/config/ConsulVariableLookup.java
@@ -1,4 +1,4 @@
-package com.wikia.pandora.core.consul.config;
+package com.wikia.dropwizard.consul.config;
 
 import com.wikia.gradle.ConsulKeyValueConfig;
