+++ lib/dropwizard/consul/src/test/java/com/wikia/dropwizard/consul/bundle/ConsulWrapperTest.java
@@ -1,4 +1,4 @@
-package com.wikia.pandora.core.consul;
+package com.wikia.dropwizard.consul.bundle;
 
 import org.junit.Test;
