+++ service/example/src/main/java/com/wikia/exampleservice/configuration/ExampleServiceConfiguration.java
@@ -1,7 +1,7 @@
 package com.wikia.exampleservice.configuration;
 
-import com.wikia.pandora.core.consul.ConsulConfiguration;
-import com.wikia.pandora.core.consul.ProvidesConsulConfiguration;
+import com.wikia.dropwizard.consul.bundle.ConsulConfiguration;
+import com.wikia.dropwizard.consul.bundle.ProvidesConsulConfiguration;
 
 import org.hibernate.validator.constraints.NotEmpty;
 
@@ -9,7 +9,8 @@ import javax.validation.constraints.NotNull;
 
 import io.dropwizard.Configuration;
 
-public class ExampleServiceConfiguration extends Configuration implements ProvidesConsulConfiguration {
+public class ExampleServiceConfiguration extends Configuration implements
+                                                               ProvidesConsulConfiguration {
 
   @NotEmpty
   @NotNull
